Added new example_10
