+++ docs/examples/0_transport/example_03_linux_socketcan.cpp
@@ -1,5 +1,5 @@
 /// @file
-/// Example of creating a libcyphal node in your project using posix UDP sockets and transport.
+/// Example of creating a libcyphal node in your project using posix SOCKETCAN media and CAN transport.
 ///
 /// @copyright
 /// Copyright (C) OpenCyphal Development Team  <opencyphal.org>
@@ -7,14 +7,14 @@
 /// SPDX-License-Identifier: MIT
 ///
 
-#include "platform/common_helpers.hpp"
-#include "platform/node_helpers.hpp"
-#include "platform/posix/posix_single_threaded_executor.hpp"
-#include "platform/posix/udp/udp_media.hpp"
-#include "platform/tracking_memory_resource.hpp"
+#include "../platform/common_helpers.hpp"
+#include "../platform/linux/can/can_media.hpp"
+#include "../platform/linux/epoll_single_threaded_executor.hpp"
+#include "../platform/node_helpers.hpp"
+#include "../platform/tracking_memory_resource.hpp"
 
+#include <libcyphal/transport/can/can_transport.hpp>
 #include <libcyphal/transport/types.hpp>
-#include <libcyphal/transport/udp/udp_transport.hpp>
 #include <libcyphal/types.hpp>
 
 #include <gmock/gmock.h>
@@ -23,6 +23,7 @@
 #include <chrono>
 #include <cstdint>
 #include <cstdlib>
+#include <cstring>
 #include <string>
 #include <vector>
 
@@ -31,11 +32,11 @@ namespace
 
 using namespace example::platform;          // NOLINT This our main concern here in this test.
 using namespace libcyphal::transport;       // NOLINT This our main concern here in this test.
-using namespace libcyphal::transport::udp;  // NOLINT This our main concern here in this test.
+using namespace libcyphal::transport::can;  // NOLINT This our main concern here in this test.
 
 using Duration        = libcyphal::Duration;
 using TimePoint       = libcyphal::TimePoint;
-using UdpTransportPtr = libcyphal::UniquePtr<IUdpTransport>;
+using CanTransportPtr = libcyphal::UniquePtr<ICanTransport>;
 
 // https://github.com/llvm/llvm-project/issues/53444
 // NOLINTBEGIN(misc-unused-using-decls, misc-include-cleaner)
@@ -43,11 +44,14 @@ using std::literals::chrono_literals::operator""s;
 using std::literals::chrono_literals::operator""ms;
 // NOLINTEND(misc-unused-using-decls, misc-include-cleaner)
 
+using testing::_;
+using testing::Eq;
 using testing::IsEmpty;
+using testing::VariantWith;
 
 // NOLINTBEGIN(cppcoreguidelines-avoid-magic-numbers, readability-magic-numbers)
 
-class Example_02_PosixUdpTransport : public testing::Test
+class Example_03_LinuxSocketCanTransport : public testing::Test
 {
 protected:
     void SetUp() override
@@ -62,8 +66,8 @@ protected:
         {
             local_node_id_ = static_cast<NodeId>(std::stoul(node_id_str));
         }
-        // Space separated list of interface addresses, like "127.0.0.1 192.168.1.162". Default is "127.0.0.1".
-        if (const auto* const iface_addresses_str = std::getenv("CYPHAL__UDP__IFACE"))
+        // Space separated list of interface addresses, like "slcan0 slcan1". Default is "vcan0".
+        if (const auto* const iface_addresses_str = std::getenv("CYPHAL__CAN__IFACE"))
         {
             iface_addresses_ = CommonHelpers::splitInterfaceAddresses(iface_addresses_str);
         }
@@ -73,8 +77,6 @@ protected:
 
     void TearDown() override
     {
-        executor_.releaseTemporaryResources();
-
         EXPECT_THAT(mr_.allocated_bytes, 0);
         EXPECT_THAT(mr_.allocations, IsEmpty());
         EXPECT_THAT(mr_.total_allocated_bytes, mr_.total_deallocated_bytes);
@@ -85,33 +87,36 @@ protected:
 
     struct State
     {
-        posix::UdpMedia::Collection media_collection_;
-        UdpTransportPtr             transport_;
+        Linux::CanMedia::Collection media_collection_;
+        CanTransportPtr             transport_;
         NodeHelpers::Heartbeat      heartbeat_;
         NodeHelpers::GetInfo        get_info_;
 
     };  // State
 
-    TrackingMemoryResource            mr_;
-    posix::PollSingleThreadedExecutor executor_{mr_};
-    TimePoint                         startup_time_{};
-    NodeId                            local_node_id_{42};
-    Duration                          run_duration_{10s};
-    std::vector<std::string>          iface_addresses_{"127.0.0.1"};
+    example::platform::TrackingMemoryResource             mr_;
+    example::platform::Linux::EpollSingleThreadedExecutor executor_;
+    TimePoint                                             startup_time_{};
+    NodeId                                                local_node_id_{42};
+    Duration                                              run_duration_{10s};
+    std::vector<std::string>                              iface_addresses_{"vcan0"};
     // NOLINTEND
 
-};  // Example_02_PosixUdpTransport
+};  // Example_03_LinuxSocketCanTransport
 
 // MARK: - Tests:
 
-TEST_F(Example_02_PosixUdpTransport, heartbeat_and_getInfo)
+TEST_F(Example_03_LinuxSocketCanTransport, heartbeat_and_getInfo)
 {
     State state;
 
-    // Make UDP transport with collection of media.
+    // Make CAN transport with collection of media.
     //
-    state.media_collection_.make(mr_, executor_, iface_addresses_);
-    CommonHelpers::Udp::makeTransport(state, mr_, executor_, local_node_id_);
+    if (!state.media_collection_.make(executor_, iface_addresses_))
+    {
+        GTEST_SKIP();
+    }
+    CommonHelpers::Can::makeTransport(state, mr_, executor_, local_node_id_);
 
     // Publish/Subscribe heartbeats.
     state.heartbeat_.makeTxSession(*state.transport_, executor_, startup_time_);
@@ -121,7 +126,7 @@ TEST_F(Example_02_PosixUdpTransport, heartbeat_and_getInfo)
     });
 
     // Bring up 'GetInfo' server.
-    state.get_info_.setName("org.opencyphal.example_02_posix_udp_transport");
+    state.get_info_.setName("org.opencyphal.example_03_linux_socketcan_transport");
     state.get_info_.makeRxSession(*state.transport_);
     state.get_info_.makeTxSession(*state.transport_);
