+++ docs/examples/CMakeLists.txt
@@ -68,7 +68,7 @@ if(IS_LINUX)
 endif()
 message(STATUS "EXAMPLES_PLATFORM_LIBS=${EXAMPLES_PLATFORM_LIBS}")
 
-file(GLOB NATIVE_EXAMPLES
+file(GLOB_RECURSE NATIVE_EXAMPLES
     LIST_DIRECTORIES false
     CONFIGURE_DEPENDS
     RELATIVE ${CMAKE_CURRENT_SOURCE_DIR}
