+++ docs/examples/platform/posix/udp/udp_media.hpp
@@ -14,8 +14,9 @@
 #include <libcyphal/transport/udp/media.hpp>
 #include <libcyphal/transport/udp/tx_rx_sockets.hpp>
 
-#include <utility>
 #include <string>
+#include <utility>
+#include <vector>
 
 namespace example
 {
